# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Srulik's lab API
  version: 0.1.0
  description: >
    HTTP API for the Srulik's lab backend. This spec is the single source of truth for
    the generated API reference.
servers:
  - url: /
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      additionalProperties: true
      properties:
        error: { type: string }
        code: { type: string }
...
    EvidenceDraft:
      type: object
      required: [content, updatedAt]
      properties:
        content: { type: string }
        updatedAt: { type: string, nullable: true, description: "ISO timestamp" }
    EvidenceDraftUpdate:
      type: object
      required: [content]
      properties:
        content: { type: string }
    EvidenceSubmitRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
    EvidenceSubmitResponse:
      type: object
      additionalProperties: true
...
paths:
...
  /api/evidence-draft:
    get:
      tags: [Evidence]
      summary: Get the current evidence draft.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [content, updatedAt]
                properties:
                  content: { type: string }
                  updatedAt: { type: string, nullable: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [Evidence]
      summary: Update the current evidence draft.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvidenceDraftUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceDraft" }
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/evidence-submit:
    post:
      tags: [Evidence]
      summary: Submit evidence for ingestion/analysis.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvidenceSubmitRequest" }
      responses:
        "202":
          description: Accepted (queued)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceSubmitResponse" }
        "400":
          description: Invalid request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/evidence-submissions:
    get:
      tags: [Evidence]
      summary: List recent evidence submissions.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/evidence-submissions/latest:
    get:
      tags: [Evidence]
      summary: Get the latest submission (if any).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/evidence-submissions/{id}:
    get:
      tags: [Evidence]
      summary: Get a submission by id.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...